+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added module `mock` to support part of the interface offered by `pytest-mock`
+
 ### Changed
 
 ### Removed